+++ src/components/CurrencyInputPanel/index.tsx
@@ -1,5 +1,5 @@
 import { Pair } from '@kyberswap/ks-sdk-classic'
-import { ChainId, Currency } from '@kyberswap/ks-sdk-core'
+import { Currency } from '@kyberswap/ks-sdk-core'
 import { Trans } from '@lingui/macro'
 import { darken, lighten, rgba } from 'polished'
 import React, { ReactNode, useCallback, useEffect, useRef, useState } from 'react'
@@ -9,11 +9,7 @@ import styled from 'styled-components'
 import { ReactComponent as DropdownSVG } from 'assets/svg/down.svg'
 import Wallet from 'components/Icons/Wallet'
 import { RowFixed } from 'components/Row'
-import { useTokenBalanceOfAnotherChain } from 'hooks/bridge'
 import useTheme from 'hooks/useTheme'
-import SelectNetwork from 'pages/Bridge/SelectNetwork'
-import { useBridgeState } from 'state/bridge/hooks'
-import { WrappedTokenInfo } from 'state/lists/wrappedTokenInfo'
 import { useCurrencyConvertedToNative } from 'utils/dmm'
 import { shortString } from 'utils/string'
 
@@ -25,9 +21,9 @@ import Card from '../Card'
 import CurrencyLogo from '../CurrencyLogo'
 import DoubleCurrencyLogo from '../DoubleLogo'
 import { Input as NumericalInput } from '../NumericalInput'
-import CurrencySearchModal, { CurrencySearchModalBridge } from '../SearchModal/CurrencySearchModal'
+import CurrencySearchModal from '../SearchModal/CurrencySearchModal'
 
-const InputRow = styled.div`
+export const InputRow = styled.div`
   ${({ theme }) => theme.flexRowNoWrap}
   align-items: center;
 `
@@ -40,7 +36,7 @@ const StyledSwitchIcon = styled(SwitchIcon)<{ selected: boolean }>`
   }
 `
 
-const CurrencySelect = styled.button<{ selected: boolean; hideInput?: boolean }>`
+export const CurrencySelect = styled.button<{ selected: boolean; hideInput?: boolean }>`
   align-items: center;
   height: ${({ hideInput }) => (hideInput ? '2.5rem' : 'unset')};
   width: ${({ hideInput }) => (hideInput ? '100%' : 'initial')};
@@ -70,13 +66,13 @@ const CurrencySelect = styled.button<{ selected: boolean; hideInput?: boolean }>
   }
 `
 
-const Aligner = styled.span`
+export const Aligner = styled.span`
   display: flex;
   align-items: center;
   justify-content: space-between;
 `
 
-const InputPanel = styled.div<{ hideInput?: boolean }>`
+export const InputPanel = styled.div<{ hideInput?: boolean }>`
   ${({ theme }) => theme.flexColumnNoWrap}
   position: relative;
   border-radius: ${({ hideInput }) => (hideInput ? '8px' : '20px')};
@@ -96,14 +92,14 @@ const FixedContainer = styled.div`
   z-index: 2;
 `
 
-const Container = styled.div<{ selected: boolean; hideInput: boolean; error?: boolean }>`
+export const Container = styled.div<{ selected: boolean; hideInput: boolean; error?: boolean }>`
   border-radius: 16px;
   background-color: ${({ theme, hideInput }) => (hideInput ? 'transparent' : theme.buttonBlack)};
   padding: ${({ hideInput }) => (hideInput ? 0 : '0.75rem')};
   border: ${({ error, theme }) => (error ? `1px solid ${theme.red}` : 'none')};
 `
 
-const StyledTokenName = styled.span<{ active?: boolean; fontSize?: string }>`
+export const StyledTokenName = styled.span<{ active?: boolean; fontSize?: string }>`
   margin-left: 0.5rem;
   font-size: ${({ active, fontSize }) => (fontSize ? fontSize : active ? '20px' : '16px')};
   overflow: hidden;
@@ -356,117 +352,6 @@ export default function CurrencyInputPanel({
             showCommonBases={showCommonBases}
           />
         )}
-      </InputPanel>
-    </div>
-  )
-}
-
-interface CurrencyInputPanelBridgeProps {
-  value: string
-  error?: boolean
-  onUserInput?: (value: string) => void
-  onMax?: () => void
-  onCurrencySelect: (currency: WrappedTokenInfo) => void
-  id: string
-  isOutput?: boolean
-  onSelectNetwork: (chain: ChainId) => void
-  chainIds: ChainId[]
-  selectedChainId: ChainId | undefined
-}
-export function CurrencyInputPanelBridge({
-  error,
-  value,
-  onUserInput = (value: string) => {
-    //
-  },
-  onSelectNetwork,
-  chainIds,
-  onMax,
-  selectedChainId,
-  onCurrencySelect,
-  isOutput = false,
-  id,
-}: CurrencyInputPanelBridgeProps) {
-  const [modalOpen, setModalOpen] = useState(false)
-  const { chainId, account } = useActiveWeb3React()
-  const [{ currencyIn, currencyOut, listTokenOut, chainIdOut, loadingToken }] = useBridgeState()
-  const currency = isOutput ? currencyOut : currencyIn
-  const selectedCurrencyBalance = useCurrencyBalance(account ?? undefined, isOutput ? undefined : currency ?? undefined)
-  const balanceRef = useRef(selectedCurrencyBalance?.toSignificant(10))
-  const destBalance = useTokenBalanceOfAnotherChain(chainIdOut, isOutput ? currency : undefined)
-
-  useEffect(() => {
-    balanceRef.current = undefined
-  }, [chainId])
-
-  // Keep previous value of balance if rpc node was down
-  useEffect(() => {
-    if (!!selectedCurrencyBalance) balanceRef.current = selectedCurrencyBalance.toSignificant(10)
-    if (!currency || !account) balanceRef.current = '0'
-  }, [selectedCurrencyBalance, currency, account])
-
-  const theme = useTheme()
-
-  const handleDismissSearch = useCallback(() => {
-    setModalOpen(false)
-  }, [setModalOpen])
-
-  const disabledSelect = listTokenOut.length === 1 && isOutput
-  const formatDestBalance = parseFloat(destBalance) ? parseFloat(destBalance)?.toFixed(10) : 0
-  return (
-    <div style={{ width: '100%' }}>
-      <InputPanel id={id}>
-        <Container hideInput={false} selected={false} error={error}>
-          <Flex justifyContent="space-between" fontSize="12px" marginBottom="12px" alignItems="center">
-            <SelectNetwork chainIds={chainIds} onSelectNetwork={onSelectNetwork} selectedChainId={selectedChainId} />
-            <Flex
-              onClick={() => onMax && onMax()}
-              style={{ cursor: onMax ? 'pointer' : undefined }}
-              alignItems="center"
-            >
-              <Wallet color={theme.subText} />
-              <Text fontWeight={500} color={theme.subText} marginLeft="4px">
-                {isOutput ? formatDestBalance : selectedCurrencyBalance?.toSignificant(10) || balanceRef.current || 0}
-              </Text>
-            </Flex>
-          </Flex>
-          <InputRow>
-            <NumericalInput
-              error={error}
-              className="token-amount-input"
-              value={value}
-              disabled={isOutput}
-              onUserInput={onUserInput}
-            />
-
-            <CurrencySelect
-              selected={!!currency}
-              className="open-currency-select-button"
-              onClick={() => !disabledSelect && !loadingToken && setModalOpen(true)}
-              style={{ cursor: disabledSelect ? 'default' : 'pointer', paddingRight: disabledSelect ? '8px' : 0 }}
-            >
-              <Aligner>
-                <RowFixed>
-                  {currency && <CurrencyLogo currency={currency} size={'20px'} />}
-                  <StyledTokenName
-                    className="token-symbol-container"
-                    active={Boolean(currency?.symbol)}
-                    style={{ paddingRight: 0 }}
-                  >
-                    {currency?.symbol || (loadingToken ? <Trans>Loading tokens</Trans> : <Trans>Select a token</Trans>)}
-                  </StyledTokenName>
-                </RowFixed>
-                {disabledSelect ? <div /> : <DropdownSVG />}
-              </Aligner>
-            </CurrencySelect>
-          </InputRow>
-        </Container>
-        <CurrencySearchModalBridge
-          isOutput={isOutput}
-          isOpen={modalOpen}
-          onDismiss={handleDismissSearch}
-          onCurrencySelect={onCurrencySelect}
-        />
       </InputPanel>
     </div>
   )
