+++ src/pages/Bridge/SwapForm.tsx
@@ -11,7 +11,7 @@ import MultichainLogoDark from 'assets/images/multichain_black.png'
 import MultichainLogoLight from 'assets/images/multichain_white.png'
 import { ReactComponent as ArrowUp } from 'assets/svg/arrow_up.svg'
 import { ButtonConfirmed, ButtonError, ButtonLight } from 'components/Button'
-import { CurrencyInputPanelBridge } from 'components/CurrencyInputPanel'
+import CurrencyInputPanelBridge from 'components/CurrencyInputPanel/CurrencyInputPanelBridge'
 import Loader from 'components/Loader'
 import ProgressSteps from 'components/ProgressSteps'
 import { AutoRow, RowBetween } from 'components/Row'
